# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Social Media application",
        "title": "Social Media Application API"
    },
    "host": "localhost:3000",
    "securityDefinitions": {
        "jwt": {}
    },
    "paths": {
...
        "/api/user/sign-in": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "User login to get token",
                "operationId": "userLogin",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User Credentials",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "required": true
                                },
                                "password": {
                                    "type": "string",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Successful"
                    },
                    "400": {
                        "description": "Please enter email/password"
                    },
                    "401": {
                        "description": "Invalid email or passsword!"
                    }
                }
            }
        },
...
        "/api/user/get-details/{userId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User Details",
                "description": "To get user details",
                "operationId": "getUserDetails",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The id of the user.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find user details."
                    },
                    "401": {
                        "description": "User not found! register yourself now!!"
                    }
                }
            }
        },
        "/api/user/admin/get-all-user-details": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All user details",
                "description": "To get all user details",
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all user details."
                    }
                }
            }
        },
        "/api/user/update-details/{userId}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user profile",
                "description": "",
                "operationId": "updateUserProfile",

                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The id of the user.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "role",
                        "in": "formData",
                        "type": "string",
                        "description": "Values are [admin or user]",
                        "required": true
                    },
                    {
                        "name": "profileImgUrl",
                        "in": "formData",
                        "type": "file",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User details updated successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
...
    },
...
}